Extract TimeLogProcessor module
