+++ lib/line-builders.js
@@ -5,16 +5,6 @@ const {
   prettifyObject
 } = require('./utils')
 
-function appendTime (lineParts, { prettified }) {
-  const { prettifiedTime } = prettified
-  if (prettifiedTime) {
-    if (lineParts.length > 0) {
-      lineParts.push(' ')
-    }
-    lineParts.push(prettifiedTime)
-  }
-}
-
 function appendEOL (lineParts, { EOL }) {
   if (lineParts.length > 0) {
     lineParts.push(EOL)
@@ -47,7 +37,6 @@ function appendObjectOrError (lineParts, { log, EOL, IDENT, errorLikeObjectKeys,
 }
 
 const builtInLineBuilders = {
-  appendTime,
   appendEOL,
   appendObjectOrError
 }
