+++ lib/log-processors.js
@@ -1,4 +1,3 @@
-const { prettifyTime } = require('./utils')
 const { builtInLineBuilders } = require('./line-builders')
 const LogProcessor = require('./LogProcessor')
 const JsonLogProcessor = require('./processors/JsonLogProcessor')
@@ -9,9 +8,9 @@ const LevelLogProcessor = require('./processors/LevelLogProcessor')
 const MessageLogProcessor = require('./processors/MessageLogProcessor')
 const MetadataLogProcessor = require('./processors/MetadataLogProcessor')
 const SemicolonLogProcessor = require('./processors/SemicolonLogProcessor')
+const TimeLogProcessor = require('./processors/TimeLogProcessor')
 
 const {
-  appendTime,
   appendEOL,
   appendObjectOrError
 } = builtInLineBuilders
@@ -28,16 +27,6 @@ function createLogProcessor (definition) {
     return new builtInLogProcessors[definition]()
   } else {
     return definition
-  }
-}
-
-class TimeLogProcessor extends LogProcessor {
-  parse (input, context) {
-    return prettifyTime(input, context)
-  }
-
-  build (lineParts, context) {
-    return appendTime(lineParts, context)
   }
 }
